Combine context, provider and hook in single file. Remove appWrapper and use provider at topmost level
