--- packages/venia-concept/src/components/App/appWrapper.js
@@ -1,18 +1,0 @@
-import React from 'react';
-import App from './container';
-
-import { WindowSizeContextProvider } from './WindowSizeContext';
-
-/**
- * A functional wrapper so that we can provide context to the entire app
- * without having to rewrite the App file itself.
- */
-function AppWrapper() {
-    return (
-        <WindowSizeContextProvider>
-            <App />
-        </WindowSizeContextProvider>
-    );
-}
-
-export default AppWrapper;
